# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Self-hosted note-taking application API. Health probes are available at root paths `/livez` and `/readyz` (outside `/api/v1`).",
        "title": "Jot API",
        "contact": {
            "name": "Jot Project",
            "url": "https://github.com/hanzei/jot"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/notes/import": {
            "post": {
                "description": "jot_json takes a Jot export bundle (.zip, as produced by GET /notes/export) of up to 1 GiB; a bare JSON file from an older export is rejected. Notes are imported all-or-nothing; an image that is missing or invalid is left off its note and listed in errors. google_keep takes a Takeout .zip or a single .json file of up to 32 MiB.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Import notes from a supported export format",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Import format: jot_json or google_keep (requires file); usememos (requires url and token)",
                        "name": "import_type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Export file (required when import_type is jot_json or google_keep)",
                        "name": "file",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Memos instance URL (required when import_type is usememos)",
                        "name": "url",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Memos API token (required when import_type is usememos)",
                        "name": "token",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ImportResponse"
                        }
                    },
                    "400": {
                        "description": "bad request, or not a readable export (invalid_import_file)",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "file too large",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "export exceeds a size or file-count limit",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "CookieAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "apierr.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code is stable and machine-readable; clients switch on it.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/apierr.Code"
                        }
                    ],
                    "example": "not_found"
                },
                "message": {
                    "description": "Message is human-readable English, for logs and as a fallback. Masked\nto \"internal server error\" for 5xx responses.",
                    "type": "string",
                    "example": "note not found"
                }
            }
        },
        "apierr.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/apierr.ErrorDetail"
                }
            }
        },
...
        "handlers.ImportResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "imported": {
                    "type": "integer"
                },
                "skipped": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "jot_session",
            "in": "header"
        }
    }
}